# backend-go/docs/openapi.yaml
openapi: 3.0.0
info:
  title: SWSE Architect API
  description: API for managing SWSE ships, libraries, hangars, and configurations.
  version: 2.0.0
servers:
  - url: http://localhost:8787
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Resource:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [ship, library, hangar, config]
        data:
          type: object
          description: Flexible JSON data depending on type
        visibility:
          type: string
          enum: [private, group, public]
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
        _links:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Link'
    Link:
      type: object
      properties:
        href:
          type: string
        rel:
          type: string
        method:
          type: string
    CreateResourceRequest:
      type: object
      required:
        - name
        - visibility
        - data
      properties:
        name:
          type: string
        visibility:
          type: string
          enum: [private, group, public]
        data:
          type: object
...
paths:
...
  /ships:
    get:
      summary: List Ships
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: List of ships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resource'
    post:
      summary: Create Ship
      security: [{bearerAuth: []}]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRequest'
      responses:
        '200':
          description: Ship created
...
  /libraries:
    get:
      summary: List Libraries
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: List of libraries
    post:
      summary: Create Library
      security: [{bearerAuth: []}]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRequest'
      responses:
        '200':
          description: Library created
...
  /hangars:
    get:
      summary: List Hangars
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: List of hangars
    post:
      summary: Create Hangar
      security: [{bearerAuth: []}]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRequest'
      responses:
        '200':
          description: Hangar created
...
  /configurations:
    get:
      summary: List Configurations
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: List of configurations
    post:
      summary: Create Configuration
      security: [{bearerAuth: []}]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRequest'
      responses:
        '200':
          description: Configuration created
...